requirements.txt
Flask==2.0.1
line-bot-sdk==1.19.0
psycopg2-binary==2.9.1
-e git+https://github.com/s1see/bitcoinlib.git#egg=bitcoinlib-0.1.0